cfs: bandwidth throttling
Merge a (quite old) version of the CFS bandwidth throttling patch (by
Bharata B Rao and Paul Turner).

Even if the patch is quite old it is enough for me to do some tests with
CPU throttling. The idea is to throttle background apps in order to
increase battery life.

If it works we can update the patch later, and merge a newer version.
For now, all the required functionalities are provided by this patch.

Signed-off-by: Andrea Righi <andrea@betterlinux.com>