sched: set sched_migration_cost to max
Raise the the migration barrier of the system to the max level. We can
tolerate to have a busy cpu and an almost idle cpu. In terms of power
consumption this is always better than having both cpus half-busy.

If we have a lot of short-lived tasks the load balancing will be
automatically done when new threads will be scheduled.

This also an aggressive policy to avoid cache misses, because all tasks
are always considered cache-hot (and so not eligible for migration).

Signed-off-by: Andrea Righi <andrea@betterlinux.com>