pm-hotplug: reduce unnecessary calls to the dynamic hotplug workqueue
Optimize the pm-hotplug governor, used to enable/disable the other cpu
at runtime depending on the system load.

Before this change the hotplug check was performed periodically every
0.5 sec using a delayed work, also when the system is totally idle.

Now we use a quite ugly solution, hooking into the scheduler an explicit
call to fire the hotplug check only when the amount of running tasks is
greater than 2 and the runqueue load reaches a certain threshold
(determined heuristicaly to 8192).

This is a very coarse grained estimation, but it seems to work very well
with many typical workloads: the other cpu (cpu1) is powered on very
quickly when the cpu0 is getting busy, and when the system is totally
idle we can save a lot of unnecessary wakeups, saving battery power.

Signed-off-by: Andrea Righi <andrea@betterlinux.com>